requirements.txt
python-telegram-bot==20.5
selenium==4.12.0
webdriver-manager==4.0.0
python-dotenv==1.0.0
colorama==0.4.6